codec: added 7 prev versions
Now the compressor and decompressor will cache the 7 most recent
distinct versions observed during this function call.

This system means the codec requires no knowledge of the blockchain
except the previous block header to still attain high compression and
decompression.

Marginal extra optimisation is still possible if compress and decompress
both took an optional argument with the previous 7 distinct versions
with which to pre-populate the prev_version deque before beginning.